requirements.txt
fastapi~=0.95.0
fastapi-sqlalchemy
uvicorn
psycopg2
starlette~=0.26.1
requests==2.28.2
httpx==0.24.0
SQLAlchemy~=2.0.7
pydantic~=1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.27.0
python-multipart
bcrypt
pytz